live-tts/core/gemini_server/requirements-gemini.txt
# requirements.txt for Gemini Server
fastapi
uvicorn
google-generativeai